move `focus.ts` to `arrows/focus.ts`
